+++ src/routes/admin/+page.server.ts
@@ -5,10 +5,8 @@ export const load: PageServerLoad = async ({ locals }) => {
 	const { user } = locals
 	const authorized = ['admin']
 	if (!user || !authorized.includes(user.role)) {
-		throw redirect(302, '/login?referrer=/admin')
+		redirect(302, '/login?referrer=/admin');
 	}
 
-	return {
-		message: 'Admin-only content from server.'
-	}
+	return { message: 'Admin-only content from server.' }
 }