+++ src/routes/register/+page.server.ts
@@ -4,7 +4,7 @@ import type { PageServerLoad } from './$types'
 export const load: PageServerLoad = ({ locals }) => {
 	const { user } = locals
 	if (user) { // Redirect to home if user is logged in already
-		throw redirect(302, '/')
+		redirect(302, '/');
 	}
 	return {}
 }